Change the hub's position information in thrusterDynamicEffector
Same module changes as before. In this case, the unit test doesn't run the simulation, and instead calls individual module functions. However, the position property is only created by a function within the spacecraft module, which wasn't being called. Therefore, we manually create the property and set it to zero so the code doesn't seg faults.