Separate excluded targets and their dependencies
Previously the custom.cmake file was executed with include() prior
to the definition of the compilation target with swig_add_library(). Splitting
the exclusion options from the custom.cmake executions, allows for the
custom.cmake to have access to the compilation target and define includes
and libraries for the target. As a result build dependencies and config can
be isolated and localized to the module's custom.cmake, decreasing
coupling and increasing compilation unit cohesion.