fix where conan version is checked
With conan 2.x the imports were failing early in the conanfile.py script before we checked for the version number.
Now the version number is checked first thing before conan sub-packcages are loaded that might
not be in conan 2.x.  This way the user is informed right away if the installed conan version is
not suitable for Basilisk.