fix tolerance of unit test
with Xcode 14.3 the finite accuracy round off behavior changed slightly.  This caused this unit test to fail as it was testing down to nano-Newton accuracy (10^-9) even though the truth values were 10^2 to 10^3 in size.  I did a deep dive to see how the values between Xcode 14.2 and 14.3 changed, and it came down to very subtle finite precision round off differences with particular sensitive evaluations.  The tolerance is now changed to micro-Newton absolute level (10^-6) which is much more reasonable for the unit test.