Remove pass through on bool type
Prior, if one has a message payload type with a member of bool type,
then this catches that bool type and returns an empty dict rather than
the expected array of bools. This adjustment ensures that bools are
caught in the final else block along with all other common python types.