Refactor lambertPlanner and lambertValidator modules
The conditions for BSK_ERROR messages for mu and the transfer time (time between maneuver time and final time)
are changed to accept mu and transfer time being equal to 0. This prevents error messages in BSK-Sims when the
Lambert modules are created (automatically) but not configured (in case they are not used).