Change logic for first time step
The first time the module is called, there is no information on the control period.
In the past, the module would simply turn all thrusters on or off depending on baseThrustState,
which could lead to undesired behavior because the remaining logic of the module is skipped
(which takes into account COM offsets via the THRArrayCmdForceMsg).
The only difference between the first and all other time steps is the unknown control period,
so now the control period is simply set to 2 seconds for the first time step, while still
executing the rest of the module.