Remove unused code
The axisChange method wasn't actually testing anything. Inspecting the git history
show that it is a relic from the initial development of the tank effector. The TeX
snippet method is no longer needed as there are many third party packages
which can parse and generate test result output in a customizable formats.