Sync EventDialog responses with broadcast Vizard
Changed protobuffer configuration to have EventReply as a separate (non-nested) message, replaced with VizEventReply. This allows Vizard to only send a single copy of event responses, which is then copied in to the VizBroadcastSyncSettings message.
Updated 'vier_nc' etc. to 'ver_nc' etc. in vizInterface.cpp to reflect name change of protobuffer message.