Add thrust scaling to mass depletion
This is significant in ensuring that mass is depleted at an appropraite rate. Because of how
the thrusterStateEffector ramps down, it asymptotically approaches zero, meaning that when
the mass depletion logic checks for non-zero thrust, it will always believe the thruster to
be on during ramp down even at extremely small values. Therefore the thrust magnitude must
be scaled during mass dot calculations to ensure virtually no mass is depleted since it will
still try to compute this.