Remove redundant conditional
The conditional is redundant because the exist_ok=True is being passed
which recreates the directory even if it exists. Recreating it is inline with
the rationale that the directory, the __init__ file and its contents should be
recreated each build, rather than requiring a -clean which is a much more
significant erasure of cached build files.