Revert "Clean up stale (previously-generated, no longer needed) message files."
This reverts commit d2e16d0f9f97e505e89253908351ada581a4a0f3.

Reverted due to discussion in merge request. We prefer that the developers
are required to manually clean (using e.g. `python conanfile.py --clean`)
the build of renamed/deleted message files to maintain the same behaviour
as for other modules. This is kept here as an explicit revert commit
in order to keep the original commit available for future reference.