ensure building documentation only runs after version bump
We want the version bump action to complete first before building documentation.  If there is an error in the version bump, we still want to build the documentation. This is because merging in an external fork will cause the version number bump to fail, but we still want to build the latest documentation