use makefile on macOS by default
Only use Xcode default if the user says to not build the project and doesn't provide a generator string.
In Xcode 16 there is an issue where Xcode is not finding "python.h".  Appears to be a regression.
This change in compiling BSK does not impact users using "python conanfile.py", the terminal
output is actually easier to read when compiled with makefile, and there are no issues with not
finding "python.h".