avoid time accuracy issues in patched conic example
Before the planet messages were only written once.  As the gravity effector is differencing the current time with the time the planet state message was written, this led to loss of time accuracy warnings.  The conversion from uint64_t to double can only handle up to about 103 days.  If it was longer, than the gravity effector was warning about the time difference having a loss of accuracy.