make statistical test more robust
Some test check if an expected values within some sigmas of the true value.  As such, such tests can fail at times without there being an issue with BSK.  Now we can run these tests in a manner such that they get repeated a few times if they fail.  A persistent error would thus still appear as a failure.

This version of pytest-rerunfailures is used to ensure compatibility with python 3.8