change the behavior of adding locations to Vizard data
With Vizard 2.3.1 the locations are now retained as a dictionary list.  There is no need to keep resending the location data unless data information has changed. Vizard now expects locations to be sent one time only and the location data is retained in Vizard.  This drastically reduces the data size.  vizSupport.py still keeps a list of all locations, but the C++ vizInterface vector of locations is cleared each time it is copied into the protobuffer data stream.

Next, we need to add the ability to change location information on the fly.