fix issue with gravity body list going out of memory
The loop over the spacecraft gravity bodies was brittle and could cause a fault if the memory of the bodies was released.  This is now made more robust.

Further, we ensure that `vizMessenger` is retained in memory by attaching it to scSim.  While this routine returns the vizMessenger instance to the parent routine, not all simulation code catches this risking the vizMessenger instance memory being deleted.